Billboard Tree and Map class

+++ src/terrain/track/TrackPieceQuarterCircle2x2.h
@@ -8,7 +8,7 @@ class TerrainManager;
 class TrackPieceQuarterCircle2x2 : public TrackPiece
 {
 public:
-	TrackPieceQuarterCircle2x2(uint16_t x, uint16_t y, uint16_t z, int rot)
+	TrackPieceQuarterCircle2x2(uint16_t x, uint8_t y, uint16_t z, int rot)
 		: TrackPiece(x, y, z, rot)
 	{ }
